Moved test classes
