+++ src/test/java/eu/tsystems/mms/tic/testerra/plugins/xray/integration/EmptyMapperTest.java
@@ -22,6 +22,8 @@ package eu.tsystems.mms.tic.testerra.plugins.xray.integration;
 
 import eu.tsystems.mms.tic.testerra.plugins.xray.AbstractTest;
 import eu.tsystems.mms.tic.testerra.plugins.xray.annotation.XrayNoSync;
+import eu.tsystems.mms.tic.testerra.plugins.xray.config.XrayConfig;
+import eu.tsystems.mms.tic.testerra.plugins.xray.hook.XrayConnectorHook;
 import eu.tsystems.mms.tic.testerra.plugins.xray.mapper.xray.XrayTestExecutionImport;
 import eu.tsystems.mms.tic.testerra.plugins.xray.mapper.xray.XrayTestExecutionIssue;
 import eu.tsystems.mms.tic.testerra.plugins.xray.mapper.xray.XrayTestIssue;
@@ -54,6 +56,7 @@ public class EmptyMapperTest extends AbstractTest {
         super.prepareWebResource();
         xrayUtils = new XrayUtils(webResource);
         verifyTestRuns = MethodsAnnotatedTest.getTestRunsForVerification();
+        XrayConnectorHook.getInstance().setXrayResultsSynchronizer(null);
     }
 
     @XrayNoSync
